ClangImporter: import size_t as Int
Converging APIs to use ``Int`` as the default integer type allows users to
write fewer explicit type conversions.

Importing ``size_t`` as a signed ``Int`` type is not be a problem for 64-bit
platforms.  32-bit platforms are important, but the usecase for an
unsigned ``size_t`` on 32-bit platforms is pretty marginal, and for code
that nevertheless needs to do that there is always the option of doing a
bitcast to ``UInt`` or using C.

rdar://18949559

Swift SVN r25070

+++ stdlib/core/String.swift
@@ -300,7 +300,7 @@ public func ==(lhs: String, rhs: String) -> Bool {
     return memcmp(
       UnsafeMutablePointer(lhs._core.startASCII),
       UnsafeMutablePointer(rhs._core.startASCII),
-      UInt(rhs._core.count)) == 0
+      rhs._core.count) == 0
   }
 #if _runtime(_ObjC)
   // Note: this operation should be consistent with equality comparison of
@@ -346,7 +346,7 @@ extension String {
     let compare = memcmp(
       UnsafeMutablePointer(self._core.startASCII),
       UnsafeMutablePointer(rhs._core.startASCII),
-      min(UInt(self._core.count), UInt(rhs._core.count)))
+      min(self._core.count, rhs._core.count))
     if compare == 0 {
       return self._core.count < rhs._core.count
     } else {
